+++ RELEASE_NOTES.md
@@ -4,3 +4,4 @@
 ### Bugfixes
 
 - Fixed drag and drop of media assets ...again.
+- Fixed empty table view when a broken VPX file was scanned.